% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperhmm.R
\name{HyperHMM}
\alias{HyperHMM}
\title{Run HyperHMM on a dataset}
\arguments{
\item{obs}{(matrix) a matrix of observations}

\item{initialstates}{(matrix) optional initial states}

\item{seed}{(integer, default 1) random number seed}

\item{nboot}{(integer, default 100) number of bootstraps}

\item{fullsample}{(integer) XXX}

\item{outputinput}{(integer) XXX}
}
\value{
A fitted HyperHMM model
}
\description{
Run HyperHMM on a dataset
}
