% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypercube-plots.R
\name{plot_bubbles}
\alias{plot_bubbles}
\title{Bubble plot from fitted model}
\usage{
plot_bubbles(fitted.obj, labels = NULL, formatted = FALSE)
}
\arguments{
\item{fitted.obj}{(fitted model)}

\item{labels}{(vector of characters) optional set of labels for features in plot}

\item{formatted}{(Boolean, default FALSE) redundant, unused}
}
\value{
A ggplot output
}
\description{
Bubble plot from fitted model
}
\examples{
m = matrix(c(0,0,1,0,1,1), byrow=TRUE, ncol=3)
fitted = HyperHMM(m, nboot=0)
plot_bubbles(fitted)
}
