% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypercube-plots.R
\name{plot_hypercube}
\alias{plot_hypercube}
\title{Hypercube plot from fitted model}
\usage{
plot_hypercube(
  fitted.obj,
  use.width = TRUE,
  duplicate.offset = 0,
  lab.size = 3,
  p.size = 1,
  node.labels = TRUE,
  seg.labels = TRUE,
  threshold = 0,
  break.redundancy = 0,
  rotate.phi = FALSE
)
}
\arguments{
\item{fitted.obj}{(fitted model)}

\item{use.width}{(Boolean, default TRUE) use line width to display edge weights?}

\item{duplicate.offset}{(double, default 0) vertical offset for nodes in identical positions}

\item{lab.size}{(double, default 3) size for edge labels}

\item{p.size}{(double, default 1) point size}

\item{node.labels}{(Boolean, default TRUE) node labels, yes or no?}

\item{seg.labels}{(Boolean, default TRUE) line segment labels, yes or no?}

\item{threshold}{(double, default 0) ignore edges under a threshold in the hypercube plot}

\item{break.redundancy}{redundant, unused}

\item{rotate.phi}{(Boolean, default FALSE) rotate states out of the page (in case of trajectories bunched up near the top/bottom)}
}
\value{
A ggplot output
}
\description{
redundant? older less flexible version -- will be sparsely commented
}
\examples{
m = matrix(c(0,0,1,0,1,1), byrow=TRUE, ncol=3)
fitted = HyperHMM(m, nboot=0)
plot_hypercube(fitted)
}
