% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypercube-plots.R
\name{plot_hypercube_flux}
\alias{plot_hypercube_flux}
\title{Adapted HyperTraPS plot for HyperHMM outputs}
\usage{
plot_hypercube_flux(
  fitted.obj,
  thresh = 0.05,
  node.labels = TRUE,
  use.probability = FALSE
)
}
\arguments{
\item{fitted.obj}{(fitted model)}

\item{thresh}{(double, default 0.05) minimum flux threshold for edge plotting}

\item{node.labels}{(Boolean, default TRUE) node labels, yes or no?}

\item{use.probability}{(Boolean, default FALSE) use probability instead of flux?}
}
\value{
A ggplot output
}
\description{
Adapted HyperTraPS plot for HyperHMM outputs
}
\examples{
m = matrix(c(0,0,1,0,1,1), byrow=TRUE, ncol=3)
fitted = HyperHMM(m, nboot=0)
plot_hypercube_flux(fitted)
}
