% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypercube-plots.R
\name{plot_pfg}
\alias{plot_pfg}
\title{Plot graph of ordered pair acquisitions}
\usage{
plot_pfg(fitted.obj, pfg.layout = "matrix", curvature = 1)
}
\arguments{
\item{fitted.obj}{(fitted model)}

\item{pfg.layout}{("matrix", "tree", "other", default "matrix") how to lay out plot}

\item{curvature}{(double, default 1) curvature of edges}
}
\value{
A ggplot output
}
\description{
Plot graph of ordered pair acquisitions
}
\examples{
m = matrix(c(0,0,1,0,1,1), byrow=TRUE, ncol=3)
fitted = HyperHMM(m, nboot=0)
plot_pfg(fitted)
}
