% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypercube-plots.R
\name{plot_standard}
\alias{plot_standard}
\title{Panel plot for model output}
\usage{
plot_standard(fitted.obj, legacy = FALSE, label = "")
}
\arguments{
\item{fitted.obj}{(fitted model) XXX}

\item{legacy}{redundant, unused}

\item{label}{(character) optional title for plot}
}
\value{
A ggplot output
}
\description{
Panel plot for model output
}
\examples{
m = matrix(c(0,0,1,0,1,1), byrow=TRUE, ncol=3)
fitted = HyperHMM(m, nboot=0)
plot_standard(fitted)
}
